requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.0
httpx==0.27.2
anthropic==0.34.0
tavily-python==0.3.3
python-docx==1.1.2
python-pptx==0.6.23
supabase==2.28.0
pydantic==2.8.2
python-multipart==0.0.9
sse-starlette==2.1.3
lxml==5.3.0
PyJWT==2.9.0
